pyproject.toml
[project]
name = "titan-oanda-algo"
version = "0.1.0"
description = "Quantitative swing trading system for OANDA — ML-driven strategy discovery, VectorBT optimisation, NautilusTrader execution, and GCE deployment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "oandapyV20>=0.7.2",
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "xgboost>=2.0",
    "joblib>=1.4",
    "pyarrow>=15.0",
    "python-dotenv>=1.0",
    "backtesting>=0.3.3",
    "vectorbt>=0.26",
    "plotly>=5.20",
    "nautilus_trader>=1.200",
]

[project.optional-dependencies]
vectorbt-pro = [
    "vectorbtpro",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "D"]
ignore = ["D100", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
